server/pyproject.toml
[project]
name = "ai-app-server"
version = "0.1.0"
description = "FastAPI + LangGraph backend for AI App."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "asyncpg>=0.30.0",
  "fastapi>=0.115.0",
  "langchain-core>=0.3.0",
  "langchain-openai>=0.2.0",
  "langgraph>=0.2.60",
  "markitdown[all]>=0.1.5",
  "motor>=3.6.0",
  "pydantic-settings>=2.6.0",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.27",
  "uvicorn[standard]>=0.30.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
